clean up filters styles
