+++ src/App.css
@@ -43,34 +43,8 @@
         width: 100% !important;
     }
 
-    .filter-container {
-        flex-direction: column;
-        align-items: center;
-        width: 90%;
-        gap: 15px;
-        margin: 15px auto;
-    }
-
-    .filter-group {
-        display: flex;
-        width: 100%;
-        justify-content: space-between;
-        gap: 10px;
-    }
-
-  .filter-group > * {
-        flex: 1;
-    }
-
   .year-tooltip, .sort-tooltip {
         width: 48%;
-    }
-
-  .year-input, .sort-select {
-        width: 100%;
-        padding: 8px;
-        border-radius: 4px;
-        border: 1px solid var(--border-color);
     }
 
     .radio-group {
@@ -78,11 +52,6 @@
         justify-content: center;
         width: 100%;
         margin: 0 auto;
-    }
-
-    .game-filter-container {
-        width: 100%;
-        display: flex;
     }
 
     .game-tooltip {
@@ -134,35 +103,8 @@
         transform: scale(1.05);
     }
 
-    .filter-container {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        gap: 20px;
-        margin: 20px auto;
-        max-width: 1000px;
-    }
-
-    .filter-group {
-        display: flex;
-        gap: 10px;
-        align-items: center;
-    }
-
   .year-tooltip, .sort-tooltip {
         min-width: 120px;
-    }
-
-  .year-input, .sort-select {
-        padding: 8px;
-        border-radius: 4px;
-        border: 1px solid var(--border-color);
-        background-color: var(--surface-color);
-    }
-
-    .game-filter-container {
-        min-width: 220px;
-        display: flex;
     }
 
     .game-tooltip {
@@ -201,23 +143,6 @@
         transform: translateY(-2px);
       box-shadow: 0 4px 12px rgba(0,0,0,0.25);
     }
-}
-
-.reset-game-button {
-    background-color: var(--surface-color);
-    border: 1px solid var(--border-color);
-    border-radius: 4px;
-    margin-left: 8px;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 0 10px;
-    font-size: 18px;
-}
-
-.reset-game-button:hover {
-    background-color: var(--border-color);
 }
 
 /* Global styles for search term highlighting */
